+++ xm.h
@@ -103,7 +103,7 @@ void xm_free_context(xm_context_t*);
  * Convert xm module to libxmlized version and output as a byte array to serial. This is an
  * 'unwrapped' version of the module that can be stored in readonly memory.
  **/
-void xm_libxmize( const char* moddata, uint16_t moddata_size );
+void xm_libxmize( const char* moddata, uint32_t moddata_size );
